Separate shapes into areas, bodies and lines
